REFACTOR: makeWorkation 페이지 최적화
