+++ src/MakeWorkation/Components/NewAnswerButton.jsx
@@ -1,13 +1,16 @@
 import React from 'react'
 import styled from 'styled-components';
 
-const AnswerButton = ({text, onClick, selected}) => {
+const NewAnswerButton = ({text, onClick, selected}) => {
   return (
-    <Button onClick={onClick} selected={selected}>{text}</Button>
+    <div style={{marginBottom: '15px', marginLeft : '20px'}}>
+        <Button onClick={onClick} selected={selected}>{text}</Button>
+    </div>
+
   )
 }
 
-export default AnswerButton
+export default NewAnswerButton
 
 const Button = styled.button`
   background-color: ${props => (props.selected ? '#FFFAE9' : '#F2F2F2')};
@@ -15,7 +18,6 @@ const Button = styled.button`
   border :  ${props => (props.selected ?  '2px solid #FED29D': 'none' )};
   padding: 10px 20px;
   margin: 3px;
-  margin-right: 15px;
   cursor: pointer;
   border-radius: 4px;
   font-size: 18px;
